Taxref v17 migrate tests commands + improve reporting (#515)
* Correction commande taxref info

* Taxref v17: ajout commande test de la migration

* Déplacement données examples pour les tests

* Improve split reporting
