+++ apptax/taxonomie/commands/migrate_taxref/commands_v16.py
@@ -1,9 +1,3 @@
-from apptax.taxonomie.commands.migrate_taxref.test_commands_v16 import (
-    populate_data,
-    clean_data,
-    test_import_taxref_v16,
-)
-
 from flask import Blueprint
 
 import importlib
@@ -197,15 +191,3 @@ def import_and_format_dbc_status():
     """
     truncate_bdc_statuts()
     import_bdc_statuts_v16(logger)
-
-
-@migrate_to_v16.command()
-@with_appcontext
-def test_taxref_v16_migration():
-    populate_data()
-    try:
-        test_import_taxref_v16()
-    except AssertionError as e:
-        raise (e)
-    finally:
-        clean_data()
