+++ apptax/taxonomie/commands/migrate_taxref/queries.py
@@ -46,15 +46,16 @@ EXPORT_QUERIES_SPLIT = """
 SELECT
     t.regne, t.group1_inpn, t.group2_inpn,
     s.i_cd_ref, s.i_array_agg AS i_cd_noms, t.nom_valide AS i_nom_valid,
-    s.f_cd_ref, s.f_array_agg AS f_cd_noms, it.nom_valide AS f_nom_valid
+    s.f_cd_ref, s.f_array_agg AS f_cd_noms, it.nom_valide AS f_nom_valid,
+    cas
 FROM tmp_taxref_changes.split_analyze s
 JOIN taxonomie.taxref t
 ON t.cd_nom = s.i_cd_ref
 JOIN taxonomie.import_taxref it
 ON it.cd_nom = s.f_cd_ref
-WHERE cas = 'split';
+WHERE not cas IS NULL;
 """
 
 EXPORT_QUERIES_NB_SPLIT = """
-SELECT count(*) FROM tmp_taxref_changes.split_analyze WHERE cas = 'split'
+SELECT count(*) FROM tmp_taxref_changes.split_analyze WHERE cas ilike '%split%'
 """
