+++ apptax/tests/fixtures.py
@@ -147,66 +147,3 @@ def users(app):
         users[user.identifiant] = user
 
     return users
-
-
-# test migration taxref détection des différents cas
-data_migration_taxref_v15_to_v16 = [
-    (106344, 106344, None, "Linum suffruticosum L., 1753", None),  # cd_nom sans substition
-    (850859, 658167, 658167, "Navicula accomoda Hust., 1950", None),  # cd_nom avec substition
-    (961306, 961306, 3945, "Motacilla yarrellii Gould, 1837", "A"),  # update cd_ref
-    (459099, 459099, 6754, "Sphagnum denticulatum f. crassicladum", "A"),  # merge
-    (6754, 6754, 6754, "Sphagnum auriculatum Schimp., 1857", "A"),  # merge
-    (443766, 443766, 443766, "Dasyprocta leporina (Linnaeus, 1758)", "A"),  # merge with conflict
-    (956958, 956958, 443766, "Mus aguti Linnaeus, 1766", "B"),  # merge with conflict
-    (1900, 1900, 1898, "Alopecosa accentuata (Latreille, 1817)", "A"),  # split
-    (233868, 1900, 233868, "Alopecosa barbipes (Sundevall, 1833)", None),  # split
-]
-
-
-def populate_migration_example(taxons):
-    liste = BibListes.query.filter_by(code_liste="100").one()
-
-    theme = BibThemes.query.filter_by(nom_theme="Mon territoire").one()
-    attribut = BibAttributs(
-        nom_attribut="test",
-        label_attribut="Test",
-        liste_valeur_attribut='{"values":["A","B","C"]}',
-        obligatoire=False,
-        type_attribut="varchar(50)",
-        type_widget="select",
-        id_theme=theme.id_theme,
-        # regne="Animalia",
-        # group2_inpn="Oiseaux",
-        ordre=1,
-    )
-    db.session.add(attribut)
-
-    with db.session.begin_nested():
-        for cd_nom, cd_ref, nom_complet, attr_value in taxons:
-            # cor_nom_liste
-            nom = Taxref.query.get(cd_nom)
-            db.session.add(nom)
-            liste.noms.append(nom)
-
-            # medias
-            media = TMedias(
-                cd_ref=cd_ref,
-                titre=nom_complet,
-                url="https://upload.wikimedia.org/wikipedia/commons/f/f0/Taxa-4x35-tagskilt.jpg",
-                id_type=1,
-                is_public=True,
-                supprime=False,
-            )
-            db.session.add(media)
-
-            # attributs
-            if attr_value:
-                attr = CorTaxonAttribut(
-                    id_attribut=attribut.id_attribut, cd_ref=cd_ref, valeur_attribut=attr_value
-                )
-                db.session.add(attr)
-
-
-@pytest.fixture
-def migration_example():
-    populate_migration_example(data_migration_taxref_v15_to_v16)
